Move authorization to a module
