+++ iml-api/src/main.rs
@@ -3,6 +3,7 @@
 // license that can be found in the LICENSE file.
 
 mod action;
+mod authorization;
 mod command;
 mod error;
 mod graphql;
